src/requirements.txt
pypdf>=4.0.0
openpyxl>=3.1.0
pandas>=2.0.0
google-genai>=1.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0
